Refactor override method
